pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "icloudsync"
version = "0.1.0"
description = "CLI para sincronizar iCloud Photos (fototeca y compartidos) a disco con Docker"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
  { name = "Victor Garcia", email = "victor.garciah@icloud.com" }
]
dependencies = [
  "typer>=0.12",
  "pyicloud-ipd>=0.10.0",
  "tqdm>=4.66",
  "python-dateutil>=2.8",
  "pytz>=2024.1",
  "tenacity>=8.2",
  "PyYAML>=6.0",
  "requests>=2.31",
  "piexif>=1.1"
]

[project.scripts]
icloudsync = "icloudsync.cli:main"

